+++ waimak-app/src/main/scala/com/coxautodata/waimak/spark/app/WaimakApp.scala
@@ -9,14 +9,14 @@ import scala.reflect.runtime.universe.TypeTag
 /**
   * This is a [[SparkApp]] specifically for applications using Waimak
   *
-  * @tparam E the type of the [[Env]] implementation (must be a case class)
+  * @tparam E the type of the [[WaimakEnv]] implementation (must be a case class)
   */
-abstract class WaimakApp[E <: Env : TypeTag] extends SparkApp[E] {
+abstract class WaimakApp[E <: Env with WaimakEnv : TypeTag] extends SparkApp[E] {
 
   override protected def run(sparkSession: SparkSession, env: E): Unit = {
-    val executor = Waimak.sparkExecutor()
+    val executor = env.maxParallelActions.map(Waimak.sparkExecutor(_)).getOrElse(Waimak.sparkExecutor())
     val emptyFlow = Waimak.sparkFlow(sparkSession, env.tmpDir)
-    executor.execute(flow(emptyFlow, env))
+    executor.execute(flow(emptyFlow, env), env.errorOnUnexecutedActions)
   }
 
   def flow(emptyFlow: SparkDataFlow, env: E): SparkDataFlow