+++ waimak-app/src/test/scala/com/coxautodata/waimak/spark/app/TestMultiAppRunner.scala
@@ -121,16 +121,18 @@ object WaimakAppWithDependency extends WaimakApp[WaimakAppWithDependencyEnv] {
 case class WaimakAppNoDependencyEnv(project: String
                                     , environment: String
                                     , branch: String
-                                    , uri: String) extends HiveEnv {
+                                    , uri: String) extends HiveEnv with WaimakEnv {
   val outputPath: String = s"$basePath/output"
+  val baseDatabaseLocation: String = s"$basePath/hive"
 }
 
 case class WaimakAppWithDependencyEnv(project: String
                                       , environment: String
                                       , branch: String
                                       , uri: String
-                                      , inputPath: String) extends HiveEnv {
+                                      , inputPath: String) extends HiveEnv with WaimakEnv {
   val outputPath: String = s"$basePath/output"
+  val baseDatabaseLocation: String = s"$basePath/hive"
 }