+++ waimak-core/src/main/scala/com/coxautodata/waimak/dataflow/spark/SparkDataFlow.scala
@@ -97,6 +97,29 @@ class SparkDataFlow(info: SparkDataFlowInfo) extends DataFlow with Logging {
       }
   }
 
+  override def finaliseExecution(): Try[SparkDataFlow.this.type] = {
+    import SparkDataFlow._
+    super.finaliseExecution()
+      .map {
+        flow =>
+          (flow.tempFolder, flow.flowContext.getBoolean(REMOVE_TEMP_AFTER_EXECUTION, REMOVE_TEMP_AFTER_EXECUTION_DEFAULT)) match {
+            case (None, _) =>
+              logInfo(s"Not cleaning up temporary folder after flow execution as it is not defined")
+              flow
+            case (Some(tmp), false) =>
+              logInfo(s"Not cleaning up temporary folder [$tmp] after flow execution as [$REMOVE_TEMP_AFTER_EXECUTION] was false")
+              flow
+            case (Some(tmp), true) =>
+              logInfo(s"Cleaning up temporary folder [$tmp] as [$REMOVE_TEMP_AFTER_EXECUTION] was true")
+              flow.flowContext.fileSystem.delete(tmp, true)
+              flow
+          }
+      }
+  }
+
+  override def executor: DataFlowExecutor = info.executor
+
+  override def withExecutor(executor: DataFlowExecutor): this.type = new SparkDataFlow(info.copy(executor = executor)).asInstanceOf[this.type]
 }
 
 case class LabelCommitDefinition(basePath: String, timestampFolder: Option[String] = None, partitions: Seq[String] = Seq.empty, connection: Option[HadoopDBConnector] = None)
@@ -161,9 +184,19 @@ case class SparkDataFlowInfo(spark: SparkSession,
                              schedulingMeta: SchedulingMeta,
                              commitLabels: Map[String, LabelCommitDefinition] = Map.empty,
                              tagState: DataFlowTagState = DataFlowTagState(Set.empty, Set.empty, Map.empty),
-                             commitMeta: CommitMeta = CommitMeta.empty)
+                             commitMeta: CommitMeta = CommitMeta.empty,
+                             executor: DataFlowExecutor = Waimak.sparkExecutor())
 
 object SparkDataFlow {
+
+  import DataFlow.dataFlowParamPrefix
+
+  /**
+    * Whether to clean up the temporary folder after a flow finishes execution.
+    * Only cleans up if execution was successful.
+    */
+  val REMOVE_TEMP_AFTER_EXECUTION: String = s"$dataFlowParamPrefix.removeTempAfterExecution"
+  val REMOVE_TEMP_AFTER_EXECUTION_DEFAULT: Boolean = true
 
   def empty(spark: SparkSession): SparkDataFlow = new SparkDataFlow(SparkDataFlowInfo(spark, DataFlowEntities.empty, Seq.empty, Set.empty, None, new SchedulingMeta()))