+++ waimak-core/src/test/scala/com/coxautodata/waimak/dataflow/spark/TestSparkDataFlow.scala
@@ -184,6 +184,29 @@ class TestSparkDataFlow extends SparkAndTmpDirSpec {
 
   }
 
+  describe("execute") {
+    it("execute a flow inline") {
+      val spark = sparkSession
+      import spark.implicits._
+
+      val (executedActions, finalState) = Waimak.sparkFlow(spark)
+        .openCSV(basePath)("csv_1", "csv_2")
+        .show("csv_1")
+        .show("csv_2")
+        .withExecutor(executor)
+        .execute()
+
+      //validate executed actions
+      executedActions.size should be(4)
+      executedActions.map(a => a.description) should be(Seq("Action: read Inputs: [] Outputs: [csv_1]", "Action: read Inputs: [] Outputs: [csv_2]", "Action: show Inputs: [csv_1] Outputs: []", "Action: show Inputs: [csv_2] Outputs: []"))
+
+      finalState.actions.size should be(0) // no actions to execute
+      finalState.inputs.size should be(2)
+      finalState.inputs.getOption[Dataset[_]]("csv_1").map(_.as[TPurchase].collect()).get should be(purchases)
+      finalState.inputs.getOption[Dataset[_]]("csv_2").map(_.as[TPerson].collect()).get should be(persons)
+    }
+  }
+
   describe("sql") {
 
     it("group by single") {
@@ -747,6 +770,7 @@ class TestSparkDataFlow extends SparkAndTmpDirSpec {
 
   describe("SequentialDataFlowExecutor") {
     it("any files in staging dir should be cleaned up before any actions are executed") {
+      sparkSession.conf.set(SparkDataFlow.REMOVE_TEMP_AFTER_EXECUTION, false)
 
       val testingDir = new File(tmpDir.toUri.getPath + "/test1")
       FileUtils.forceMkdir(testingDir)
@@ -759,6 +783,7 @@ class TestSparkDataFlow extends SparkAndTmpDirSpec {
     }
 
     it("any empty staging folder should be created when a flow is executed") {
+      sparkSession.conf.set(SparkDataFlow.REMOVE_TEMP_AFTER_EXECUTION, false)
 
       val tmpDirFile = new File(tmpDir.toUri.getPath)
       tmpDirFile.getParentFile.list().toSeq should be(Seq())
@@ -793,6 +818,42 @@ class TestSparkDataFlow extends SparkAndTmpDirSpec {
       emptyFlow.flowContext.uriUsed.toString should be("file:///")
       FileSystem.get(spark.sparkContext.hadoopConfiguration).getUri.toString should be("hdfs://localhost")
     }
+  }
+
+  describe("finaliseExecution") {
+
+    it("should delete a temporary folder after execution by default"){
+      val emptyFlow = Waimak.sparkFlow(sparkSession, tmpDir.toString)
+
+      Waimak.sparkExecutor().execute(emptyFlow)
+      emptyFlow.flowContext.fileSystem.exists(tmpDir) should be (false)
+
+    }
+
+    it("should not delete a temporary folder after execution if configuration is set"){
+
+      sparkSession.conf.set("spark.waimak.dataflow.removeTempAfterExecution", false)
+      val emptyFlow = Waimak.sparkFlow(sparkSession, tmpDir.toString)
+
+      Waimak.sparkExecutor().execute(emptyFlow)
+      emptyFlow.flowContext.fileSystem.exists(tmpDir) should be (true)
+
+    }
+
+    it("should not delete a temporary folder after execution the flow fails"){
+
+      val badFlow = Waimak.sparkFlow(sparkSession, tmpDir.toString).open("bad", _ => throw new RuntimeException("bad action"))
+
+      badFlow.flowContext.getBoolean("spark.waimak.dataflow.removeTempAfterExecution", SparkDataFlow.REMOVE_TEMP_AFTER_EXECUTION_DEFAULT) should be (true)
+
+      intercept[RuntimeException] {
+        Waimak.sparkExecutor().execute(badFlow)
+      }.getCause.getMessage should be ("bad action")
+
+      badFlow.flowContext.fileSystem.exists(tmpDir) should be (true)
+
+    }
+
   }
 
   describe("mixing multiple types in flow") {