+++ waimak-storage/src/test/scala/com/coxautodata/waimak/storage/TestAuditTableFile.scala
@@ -90,6 +90,20 @@ class TestAuditTableFile extends SparkAndTmpDirSpec {
       infoData should be(Success(AuditTableInfo(tableName, Seq("id"), Map.empty, true)))
     }
 
+    it("init table then update table info") {
+      val zeroState = createADTable(tableName, createFops())
+
+      val nextState = zeroState.initNewTable().get
+
+      val infoData = zeroState.storageOps.readAuditTableInfo(basePath, tableName)
+      infoData should be(Success(AuditTableInfo(tableName, Seq("id"), Map.empty, true)))
+
+      val updatedTableInfoState = nextState.updateTableInfo(AuditTableInfo(tableName, Seq("id1", "id2"), Map.empty, false))
+
+      val updatedInfoData = zeroState.storageOps.readAuditTableInfo(basePath, tableName)
+      updatedInfoData should be(Success(AuditTableInfo(tableName, Seq("id1", "id2"), Map.empty, false)))
+    }
+
     it("init table fail") {
       val zeroState = createADTable(tableName, createFops())
 
@@ -242,7 +256,7 @@ class TestAuditTableFile extends SparkAndTmpDirSpec {
 
       val secondCompact = table_s3.compact(lastTS_4, d12h, SMALL_REGION_ROW_THRESHOLD_DEFAULT, defaultCompactionPartitioner).get
       secondCompact.regions.size should be(1)
-      secondCompact.regions(0).store_region should be("6")
+      secondCompact.regions(0).store_region should be("5")
 
       // Should be a different trashed compaction, original one would be deleted
       new File(trashBinPath.toString, tableName).list() should contain theSameElementsAs Seq(lastTS_4.getTime.toString)
@@ -250,19 +264,19 @@ class TestAuditTableFile extends SparkAndTmpDirSpec {
       // Empty compaction with previous in range should not delete trash
       val thirdCompact = secondCompact.compact(lastTS_5, d3d, SMALL_REGION_ROW_THRESHOLD_DEFAULT, defaultCompactionPartitioner).get
       thirdCompact.regions.size should be(1)
-      thirdCompact.regions(0).store_region should be("6")
+      thirdCompact.regions(0).store_region should be("5")
       new File(trashBinPath.toString, tableName).list() should contain theSameElementsAs Seq(lastTS_4.getTime.toString)
 
       // Empty compaction with previous not in range should delete trash
       val fourthCompact = thirdCompact.compact(lastTS_6, d12h, SMALL_REGION_ROW_THRESHOLD_DEFAULT, defaultCompactionPartitioner).get
       fourthCompact.regions.size should be(1)
-      fourthCompact.regions(0).store_region should be("6")
+      fourthCompact.regions(0).store_region should be("5")
       new File(trashBinPath.toString, tableName).list() should contain theSameElementsAs Seq()
 
       //Request to recompact all should force single cold region to be recompacted
       val fifthCompact = fourthCompact.compact(lastTS_7, d3d, SMALL_REGION_ROW_THRESHOLD_DEFAULT, defaultCompactionPartitioner, recompactAll = true).get
       fifthCompact.regions.size should be(1)
-      fifthCompact.regions(0).store_region should be("7")
+      fifthCompact.regions(0).store_region should be("6")
     }
 
     it("single compact into new with schema evolution on hot") {
@@ -366,10 +380,10 @@ class TestAuditTableFile extends SparkAndTmpDirSpec {
       reopenedTable.map(_.compact(lastTS_3, d3d, SMALL_REGION_ROW_THRESHOLD_DEFAULT, defaultCompactionPartitioner))
 
       val inferredCachedRegionsAfterCompaction = AuditTableFile.inferRegionsFromCache(person.storageOps, basePath, Seq("prsn"), true)
-      inferredCachedRegionsAfterCompaction should be(Seq(AuditTableRegionInfo("prsn", "cold", "5", lastTS_3, false, 11, lastTS_3)))
+      inferredCachedRegionsAfterCompaction should be(Seq(AuditTableRegionInfo("prsn", "cold", "4", lastTS_3, false, 11, lastTS_3)))
 
       val inferredAllRegionsAfterCompaction = AuditTableFile.inferRegionsWithStats(sparkSession, person.storageOps, basePath, Seq("prsn"), true)
-      inferredAllRegionsAfterCompaction should be(Seq(AuditTableRegionInfo("prsn", "cold", "5", lastTS_3, false, 11, lastTS_3)))
+      inferredAllRegionsAfterCompaction should be(Seq(AuditTableRegionInfo("prsn", "cold", "4", lastTS_3, false, 11, lastTS_3)))
 
     }
 
@@ -740,15 +754,15 @@ class TestAuditTableFile extends SparkAndTmpDirSpec {
       import spark.implicits._
       val personTable = createADTable("person", createFops(), retainHistory = false).initNewTable().get
       val personData = persons.toDS().withColumn("lastTS", lit("2018-01-01"))
-      personTable
+      val res = personTable
         .append(personData, lastUpdated(personData), lastTS_1)
         .flatMap(_._1.append(personData, lastUpdated(personData), lastTS_1))
         .flatMap(_._1.compact(lastTS_2, d3d, SMALL_REGION_ROW_THRESHOLD_DEFAULT, defaultCompactionPartitioner))
 
+      res shouldBe a[Success[_]]
       val compactedData = spark.read.parquet(personTable.tablePath.toString)
       compactedData.as[TPerson].collect() should contain theSameElementsAs (persons)
     }
-
   }
 
   describe("clearTableRegionCache") {